pyproject.toml
[project]
name = "nullmove"
version = "0.1.0"
description = "NullMove: a minimal UCI chess engine baseline"
requires-python = ">=3.10"
dependencies = [
  "python-chess==1.999",
]

[project.scripts]
nullmove = "nullmove.uci:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
